Moved the Win8 version to a WinMD component
Most code from RateReminder was moved to RateHelper, a plain old CLR
class. This allowed access to these functions from JavaScript where an
exception was being thrown when an attempt was made to instantiate the
control. Sample.Win8 was changed to SampleCS.Win8 and a new
SampleJS.Win8 was added to demonstrate how to use the functionality from
a JavaScript application.